refactor(auth): :recycle: using new DTOs adding email
