+++ frontend/src/app/features/configuration/account/account.component.html
@@ -32,12 +32,6 @@
                         <input type="text" class="form-control" id="lastName" formControlName="lastName">
                     </div>
                 </div>
-                <div class="row py-2" [formGroup]="userInfoForm">
-                    <label for="email" class="col-sm-2 col-form-label">E-Mail</label>
-                    <div class="col-sm-10">
-                        <input type="text" class="form-control" id="email" formControlName="email">
-                    </div>
-                </div>
                 <div class="row py-3">
                     <div class="col">
                         <button class="btn btn-primary" type="button" (click)="updateInfo()">Update information</button>
